requirements.txt
ib_insync>=0.9.86
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
yfinance>=0.2.36
# ta-lib>=0.4.28
pytest>=7.4.0
python-dotenv>=1.0.0
dnspython>=2.4.2
requests>=2.31.0 